Refactor MCP diagnostics helpers
